requirements.txt
Flask
Flask-SocketIO
eventlet
mysql-connector-python
kafka-python